+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Export) test results are validated again.
 - (Export) Removed expendable JSON content when exporting unlit materials without color or texture applied.
 - Primitve mode LINE_LOOP works as expected (thanks [Hexer611][Hexer611] for [#22](https://github.com/Unity-Technologies/com.unity.cloud.gltfast/pull/22)).
+- (Test) Fail export test if glTF JSON contains unexpected or misses expected properties.
 
 ## [6.7.1] - 2024-08-07